Rework MPC and the PID_PROFILE Manager to work together
Reworking the PID_PROFILE manager to be scalable and move load and save for configs into the respective controls so we have a more object oriented approach.
Furthermore rework MPC to properly interact with the profile Manager
@dalegaard while is has been working well for a few months on my machines now, please take another look at it to make sure I didnt miss anything please
